add SubscribeCurrencyPairs to remaining providers
